Refactor project settings dialog tabs into individual components
