+++ src/pages/Project/components/ComponentCard.jsx
@@ -26,7 +26,7 @@ const ComponentCard = React.memo(
     const { header } = content;
 
     const handleClick = () => {
-      toggleOpen(id, !open, 'sidebar');
+      toggleOpen(id, !open);
     };
 
     const handleDelete = () => {
