+++ src/pages/Project/dialogs/ProjectSettingsDialog.jsx
@@ -1,66 +1,32 @@
-import React, { useContext, useState, useCallback, useEffect } from 'react';
-import { DragDropContext, Droppable } from 'react-beautiful-dnd';
-import { v4 as uuid } from 'uuid';
-import { DropzoneArea } from 'material-ui-dropzone';
+import React, { useContext, useState } from 'react';
 import {
-  Typography,
-  Grid,
   useMediaQuery,
   useTheme,
   Button,
   Dialog,
   DialogActions,
   DialogContent,
-  Switch,
   Paper,
   Tabs,
   Tab
 } from '@material-ui/core';
 
-import {
-  projects,
-  getProjectIndexById,
-  AVAILABLE_SIDEBAR_COMPONENTS,
-  AVAILABLE_PAGES
-} from '../../../mocks';
+import { useStyles } from './ProjectSettingsDialogStyles';
+import { projects, getProjectIndexById } from '../../../mocks';
 import { DialogContext, ProjectContext } from '../../../contexts';
-import {
-  ComponentCard,
-  AvailableCard,
-  AvailablePage,
-  CurrentPage
-} from '../components';
-import { useStyles, getListStyle } from './ProjectSettingsDialogStyles';
+import ProjectSettingsGeneral from './ProjectSettingsGeneral';
+import ProjectSettingsPages from './ProjectSettingsPages';
+import ProjectSettingsSidebar from './ProjectSettingsSidebar';
 
 export default function ProjectSettingsDialog() {
   const classes = useStyles();
   const theme = useTheme();
   const fullScreen = useMediaQuery(theme.breakpoints.down('sm'));
-
   const { open, setOpen } = useContext(DialogContext);
-  const { project, projectId } = useContext(ProjectContext);
-  const [currentComponents, setCurrComponents] = useState([]);
-  const [currentPages, setCurrPages] = useState([]);
+  const { project, projectId, currentPages, currentComponents } = useContext(
+    ProjectContext
+  );
   const [tabValue, setTabValue] = useState(0);
-  const [toggles, setToggles] = useState({
-    requestToJoin: true,
-    ButtonName: false
-  });
-
-  useEffect(() => {
-    if (project.meta && project.sidebar_modules) {
-      const getCurrComponents = () => {
-        const newSidebarModules = project.sidebar_modules.map((el) => {
-          return { ...el, id: uuid() };
-        });
-        const newPages = [...project.meta.pages_order];
-
-        setCurrComponents(newSidebarModules);
-        setCurrPages(newPages);
-      };
-      getCurrComponents();
-    }
-  }, [open, project]);
 
   const handleClose = () => {
     setOpen(false);
@@ -68,10 +34,6 @@ export default function ProjectSettingsDialog() {
 
   const handleTabs = (event, newValue) => {
     setTabValue(newValue);
-  };
-
-  const handleToggles = (event) => {
-    setToggles({ ...toggles, [event.target.name]: event.target.checked });
   };
 
   const handleSave = () => {
@@ -90,7 +52,8 @@ export default function ProjectSettingsDialog() {
     const cleanSidebarModules = () => {
       return currentComponents.map((el) => ({
         type: el.type,
-        content: el.content
+        content: el.content,
+        id: el.id
       }));
     };
 
@@ -113,319 +76,17 @@ export default function ProjectSettingsDialog() {
     setOpen(false);
   };
 
-  const changePageTitle = (id, newTitle) => {
-    const newPages = [...currentPages];
-    const index = newPages.findIndex((el) => el.id === id);
-    if (index !== -1) {
-      newPages[index].title = newTitle;
-      setCurrPages(newPages);
-    }
-  };
-
-  const togglePagesSettings = (id, destination, settingBool) => {
-    const newPages = [...currentPages];
-    const index = newPages.findIndex((el) => el.id === id);
-
-    if (index !== -1) {
-      newPages[index][destination] = settingBool;
-      setCurrPages(newPages);
-    }
-  };
-
-  const toggleOpen = (id, newOpen, destination) => {
-    switch (destination) {
-      case 'sidebar':
-        const newCurrComponents = [...currentComponents];
-        const index = newCurrComponents.findIndex((item) => item.id === id);
-
-        if (index !== -1) {
-          newCurrComponents[index].open = newOpen;
-          setCurrComponents(newCurrComponents);
-        }
-        break;
-      case 'pages':
-        const newCurrPages = [...currentPages];
-        const pageIndex = newCurrPages.findIndex((item) => item.id === id);
-
-        if (pageIndex !== -1) {
-          newCurrPages[pageIndex].open = newOpen;
-          setCurrPages(newCurrPages);
-        }
-        break;
-      default:
-        return;
-    }
-  };
-
-  const updateContent = useCallback(
-    (id, content) => {
-      setCurrComponents((newCurrComponents) => {
-        const index = newCurrComponents.findIndex((item) => item.id === id);
-        if (index !== -1) {
-          newCurrComponents[index].content = content;
-          return newCurrComponents;
-        }
-      });
-    },
-    [setCurrComponents]
-  );
-
-  const deleteItem = useCallback(
-    (id, destination) => {
-      switch (destination) {
-        case 'sidebar':
-          setCurrComponents((newCurrComponents) => {
-            const index = newCurrComponents.findIndex((item) => item.id === id);
-            newCurrComponents.splice(index, 1);
-            return [...newCurrComponents];
-          });
-          break;
-        case 'pages':
-          setCurrPages((newCurrPages) => {
-            const index = newCurrPages.findIndex((item) => item.id === id);
-            newCurrPages.splice(index, 1);
-            return [...newCurrPages];
-          });
-          break;
-        default:
-          return;
-      }
-    },
-    [setCurrComponents, setCurrPages]
-  );
-
-  // drag and drop logic
-  const reorder = (list, startIndex, endIndex) => {
-    const result = Array.from(list);
-    const [removed] = result.splice(startIndex, 1);
-    result.splice(endIndex, 0, removed);
-
-    return result;
-  };
-
-  const copy = (source, destination, droppableSource, droppableDestination) => {
-    const sourceClone = Array.from(source);
-    const destClone = Array.from(destination);
-    const item = sourceClone[droppableSource.index];
-
-    destClone.splice(droppableDestination.index, 0, {
-      ...item,
-      id: uuid(),
-      content: { ...item.content }
-    });
-    return destClone;
-  };
-
-  const checkUniqueness = (item, destination) => {
-    // return true to disable drag component
-    if (item.unique && destination === 'pages') {
-      const result = currentPages.find((el) => el.type === item.type);
-      return Boolean(result);
-    } else if (item.unique && destination === 'sidebar') {
-      const result = currentComponents.find((el) => el.type === item.type);
-      return Boolean(result);
-    }
-    return false;
-  };
-
-  const onDragEnd = (result) => {
-    const { destination, source } = result;
-
-    if (!destination) return;
-
-    const start = source.droppableId;
-    const finish = destination.droppableId;
-
-    if (start === finish) {
-      if (start === 'current-components') {
-        setCurrComponents(
-          reorder(currentComponents, source.index, destination.index)
-        );
-        return;
-      } else if (start === 'current-pages') {
-        setCurrPages(reorder(currentPages, source.index, destination.index));
-        return;
-      }
-    }
-
-    if (start === 'available-components' && finish === 'current-components') {
-      setCurrComponents(
-        copy(
-          AVAILABLE_SIDEBAR_COMPONENTS,
-          currentComponents,
-          source,
-          destination
-        )
-      );
-      return;
-    }
-
-    if (start === 'available-pages' && finish === 'current-pages') {
-      setCurrPages(copy(AVAILABLE_PAGES, currentPages, source, destination));
-      return;
-    }
-  };
-
-  const renderToggles = () => {
-    const togglesArray = [
-      ['Show "Request to Join" button', 'requestToJoin'],
-      ['Lorem Ipsum', 'ButtonName']
-    ];
-
-    return togglesArray.map((el) => (
-      <React.Fragment key={el[1]}>
-        <Grid item xs={10}>
-          <Typography>{el[0]}</Typography>
-        </Grid>
-        <Grid item xs={2} className={classes.switch}>
-          <Switch
-            checked={toggles[el[1]]}
-            onChange={handleToggles}
-            color="primary"
-            name={el[1]}
-          />
-        </Grid>
-      </React.Fragment>
-    ));
-  };
-
   const renderContent = (tabIndex) => {
     if (tabIndex === 0) {
-      return (
-        <Grid container spacing={2} alignItems="center">
-          <DropzoneArea
-            acceptedFiles={['image/png', 'image/jpg', 'image/jpeg']}
-            dropzoneText={'Drag and drop an image here or click'}
-            filesLimit={1}
-            maxFileSize={5000000}
-            // onChange={(files) => console.log('Files:', files)}
-          />
-          {renderToggles()}
-        </Grid>
-      );
+      return <ProjectSettingsGeneral />;
     } else if (tabIndex === 1) {
-      return (
-        <Grid container spacing={3}>
-          <Grid item xs={4}>
-            <Typography variant="h5" component="h2" gutterBottom>
-              Available Pages
-            </Typography>
-            <Droppable droppableId="available-pages" isDropDisabled={true}>
-              {(provided, snapshot) => (
-                <div
-                  ref={provided.innerRef}
-                  className={classes.availableModules}
-                >
-                  {[...AVAILABLE_PAGES].map((item, index) => {
-                    const isDisabled = checkUniqueness(item, 'pages');
-                    return (
-                      <AvailablePage
-                        key={index + item.type}
-                        index={index}
-                        item={item}
-                        isDisabled={isDisabled}
-                      />
-                    );
-                  })}
-                  {provided.placeholder}
-                </div>
-              )}
-            </Droppable>
-          </Grid>
-          <Grid item xs={8}>
-            <div>
-              <Typography variant="h5" component="h2" gutterBottom>
-                Public View
-              </Typography>
-              <Droppable droppableId="current-pages">
-                {(provided, snapshot) => (
-                  <div
-                    {...provided.droppableProps}
-                    ref={provided.innerRef}
-                    style={getListStyle(snapshot.isDraggingOver)}
-                  >
-                    {currentPages.map((item, index) => (
-                      <CurrentPage
-                        key={item.type + index}
-                        item={item}
-                        index={index}
-                        toggleOpen={toggleOpen}
-                        toggleSettings={togglePagesSettings}
-                        changeTitle={changePageTitle}
-                        deleteItem={(id) => deleteItem(id, 'pages')}
-                      />
-                    ))}
-                    {provided.placeholder}
-                  </div>
-                )}
-              </Droppable>
-            </div>
-          </Grid>
-        </Grid>
-      );
+      return <ProjectSettingsPages />;
     } else if (tabIndex === 2) {
       return <></>;
     } else if (tabIndex === 3) {
       return <></>;
     } else if (tabIndex === 4) {
-      return (
-        <Grid container spacing={3}>
-          <Grid item xs={4}>
-            <Typography variant="h5" component="h2" gutterBottom>
-              Available Modules
-            </Typography>
-            <Droppable droppableId="available-components" isDropDisabled={true}>
-              {(provided, snapshot) => (
-                <div
-                  ref={provided.innerRef}
-                  className={classes.availableModules}
-                >
-                  {AVAILABLE_SIDEBAR_COMPONENTS.map((item, index) => {
-                    const isDisabled = checkUniqueness(item, 'sidebar');
-                    return (
-                      <AvailableCard
-                        key={index + item.type}
-                        index={index}
-                        item={item}
-                        isDisabled={isDisabled}
-                      />
-                    );
-                  })}
-                  {provided.placeholder}
-                </div>
-              )}
-            </Droppable>
-          </Grid>
-          <Grid item xs={8}>
-            <Typography variant="h5" component="h2" gutterBottom>
-              Current Modules
-            </Typography>
-            <Droppable droppableId="current-components">
-              {(provided, snapshot) => (
-                <div
-                  {...provided.droppableProps}
-                  ref={provided.innerRef}
-                  style={getListStyle(snapshot.isDraggingOver)}
-                >
-                  {currentComponents.map((item, index) => (
-                    <ComponentCard
-                      key={item.type + index}
-                      item={item}
-                      index={index}
-                      toggleOpen={toggleOpen}
-                      deleteItem={(id) => deleteItem(id, 'sidebar')}
-                      updateContent={(id, content) =>
-                        updateContent(id, content)
-                      }
-                    />
-                  ))}
-                  {provided.placeholder}
-                </div>
-              )}
-            </Droppable>
-          </Grid>
-        </Grid>
-      );
+      return <ProjectSettingsSidebar />;
     }
   };
 
@@ -455,9 +116,7 @@ export default function ProjectSettingsDialog() {
         </Tabs>
       </Paper>
       <DialogContent className={classes.dialogContent}>
-        <DragDropContext onDragEnd={onDragEnd}>
-          {renderContent(tabValue)}
-        </DragDropContext>
+        {renderContent(tabValue)}
       </DialogContent>
       <DialogActions>
         <Button onClick={handleClose} color="secondary" autoFocus>
